Adding code from Chapter 21: Global Variables (https://craftinginterpreters.com/global-variables.html)

+++ src/chunk.c
@@ -73,8 +73,8 @@ uint32_t getLine(const Chunk* chunk, const size_t instructionOffset) {
   }
 }
 
-size_t addConstant(Chunk* chunk, const Value value) {
+uint32_t addConstant(Chunk* chunk, const Value value) {
   writeValueArray(&chunk->constants, value);
 
-  return chunk->constants.count - 1;
+  return (uint32_t)chunk->constants.count - 1;
 }
